[NBS-37] Add zoop support and update dependencies
- Bump editions from 2021 to 2024 in Cargo.toml
- Update dependencies: rand to 0.9.2, macroquad to 0.4.14, vecmath to
1.0.0
- Refactor `engine` module to `gui::engine` for clearer separation of
concerns
- Remove old `engine/engine.rs`, move logic to `gui/engine.rs` with
zoom/camera support
- Adjust imports and references in `main.rs` and physics modules
accordingly
